requirements.txt
fastapi
uvicorn[standard]
jpype1
pydantic
python-dotenv
